inline domain logic into cognito-user-pool and remove separate domain component
